Fix most CONFIG_DEBUG_RAM_SETUP issues.
The intel/xe7501devkit is still broken, I think the (romcc) image is too big to
fit in the bootblock if CONFIG_DEBUG_RAM_SETUP is enabled. It would make sense
to convert all CPU_INTEL_SOCKET_MPGA604 to CAR, but I have no hardware to test.

Acked-by: Stefan Reinauer <stepan@coreboot.org>


git-svn-id: svn://svn.coreboot.org/coreboot/trunk@6215 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1

+++ src/northbridge/intel/e7501/raminit.c
@@ -1044,9 +1044,9 @@ static void configure_e7501_ram_addresses(const struct mem_controller
 		sz = spd_get_dimm_size(dimm_socket_address);
 
 		RAM_DEBUG_MESSAGE("dimm size =");
-		RAM_DEBUG_HEX32(sz.side1);
+		RAM_DEBUG_HEX32((u32)sz.side1);
 		RAM_DEBUG_MESSAGE(" ");
-		RAM_DEBUG_HEX32(sz.side2);
+		RAM_DEBUG_HEX32((u32)sz.side2);
 		RAM_DEBUG_MESSAGE("\n");
 
 		if (sz.side1 == 0)
